Update model based on latest requirement discussions
Affects #6 and #7

We are now going to leverage as much info from the phone contacts system
as possible, so a Person is mostly a link to a contact. This results in
a bunch of changes to the model.

Also this commit includes initial adds of the DescriptionDetails screen
for a child.